+++ src/modules/logger/components/index.ts
@@ -1,0 +1,2 @@
+export * from './console/console.component';
+export * from './console/console.component.animations';