+++ src/modules/logger/interfaces/index.ts
@@ -1,4 +1,6 @@
 export * from './loggerConfiguration.interface';
+export * from './loggerConsoleComponent.interface';
 export * from './loggerEnricher.interface';
 export * from './loggerFilter.interface';
+export * from './loggerRest.interface';
 export * from './loggerSink.interface';