+++ src/modules/logger/misc/index.ts
@@ -1,3 +1,4 @@
 export * from './providers';
+export * from './tokens';
 export * from './types';
 export * from './utils';