+++ src/modules/logger/misc/providers.ts
@@ -1,9 +1,11 @@
-import {EnvironmentProviders, FactoryProvider, makeEnvironmentProviders} from '@angular/core';
+import {ClassProvider, EnvironmentProviders, FactoryProvider, Type, ValueProvider, makeEnvironmentProviders} from '@angular/core';
 import {Action1} from '@jscrpt/common';
 
-import {LoggerConfiguration} from '../interfaces';
+import {LoggerConfiguration, LoggerRestClient} from '../interfaces';
 import {LoggerConfigurationImpl} from '../types/loggerConfiguration';
 import {LoggerOptions} from '../types';
+import {LOGGER_REST_CLIENT} from './tokens';
+import {ConsoleComponentSinkOptions, RestSinkOptions} from '../services';
 
 /**
  * Provides configuration for default built-in logger
@@ -21,6 +23,54 @@ export function provideLoggerConfig(configBuilder: Action1<LoggerConfiguration>)
         {
             provide: LoggerOptions,
             useFactory: config.buildOptions.bind(config),
+        }
+    ]);
+}
+
+/**
+ * Provides logger rest client
+ * @param type - Type that implements `LoggerRestClient`
+ */
+export function provideLoggerRestClient(type: Type<LoggerRestClient>): EnvironmentProviders
+{
+    return makeEnvironmentProviders(
+    [
+        <ClassProvider>
+        {
+            provide: LOGGER_REST_CLIENT,
+            useClass: type,
+        }
+    ]);
+}
+
+/**
+ * Provides `ConsoleComponentSinkOptions`
+ * @param options - Value of options to be provided
+ */
+export function provideConsoleComponentSinkOptions(options: ConsoleComponentSinkOptions): EnvironmentProviders
+{
+    return makeEnvironmentProviders(
+    [
+        <ValueProvider>
+        {
+            provide: ConsoleComponentSinkOptions,
+            useValue: options,
+        }
+    ]);
+}
+
+/**
+ * Provides `RestSinkOptions`
+ * @param options - Value of options to be provided
+ */
+export function provideRestSinkOptions(options: RestSinkOptions): EnvironmentProviders
+{
+    return makeEnvironmentProviders(
+    [
+        <ValueProvider>
+        {
+            provide: RestSinkOptions,
+            useValue: options,
         }
     ]);
 }