+++ src/modules/logger/services/index.ts
@@ -1,6 +1,10 @@
+export * from './console/consoleComponent.sink';
+export * from './console/consoleComponentSink.options';
 export * from './developmentConsoleSink/developmentConsole.sink';
 export * from './logger/logger.interface';
 export * from './logger/logger.service';
 export * from './logLevelEnricher/logLevelEnricher.enricher';
+export * from './rest/rest.sink';
+export * from './rest/restSink.options';
 export * from './subLoggerSink/subLogger.sink';
-// export * from './timestampEnricher/timestamp.enricher';
+export * from './timestampEnricher/timestamp.enricher';